Fix proxy upstart conf to start on reboot
The line `start on started nginx` wasn't actually doing anything
except for preventing proxy from being automatically respawned; for
example, in the event that unattended-upgrades caused a reboot. Nginx
is a SysV init process, which means it doesn't ever emit the `started
nginx` event that our proxy upstart config was waiting for.

This fix starts proxy when the filesystem has mounted, and manually
waits for nginx to be running by polling `service nginx status`. This
might not be entirely necessary, as I only observed nginx already
running by the time the proxy backend script runs, but it seems like a
worthy precaution.